use makefile to run scripts for benching
