+++ Makefile
@@ -81,3 +81,15 @@ heaptrack:
 
 flamegraph: compile-cairo compile-starknet
 	CARGO_PROFILE_RELEASE_DEBUG=true cargo flamegraph --root --bench internals
+
+bench-all:
+	cd scripts/ && ./run_bench_deploy.sh && ./run_bench_invoke.sh && ./run_bench_deploy_invoke.sh
+
+bench-deploy:
+	cd scripts/ && ./run_bench_deploy.sh
+
+bench-invoke:
+	cd scripts/ && ./run_bench_invoke.sh
+
+bench-deploy-invoke:
+	cd scripts/ && ./run_bench_deploy_invoke.sh
