+++ src/testing/mod.rs
@@ -1,4 +1,3 @@
-mod bench_integration;
 pub mod starknet_state;
 pub mod starknet_state_error;
 pub mod type_utils;
